Lab3_REST_with_postgresql/src/goal_task_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pyjwt
python-jose[cryptography]
